+++ Examples/SwiftFortuneWheelDemoiOS/SwiftFortuneWheelDemoiOS/ViewControllers/Various/VariousWheelSimple/VariousWheelSimpleViewController.swift
@@ -1,0 +1,56 @@
+//
+//  VariousWheelSimpleViewController.swift
+//  SwiftFortuneWheelDemoiOS
+//
+//  Created by Sherzod Khashimov on 7/10/20.
+//  Copyright © 2020 Sherzod Khashimov. All rights reserved.
+//
+
+import UIKit
+import SwiftFortuneWheel
+
+class VariousWheelSimpleViewController: UIViewController {
+    
+    @IBOutlet weak var centerView: UIView! {
+        didSet {
+            centerView.layer.cornerRadius = centerView.bounds.width / 2
+            centerView.layer.borderColor = CGColor.init(srgbRed: CGFloat(256), green: CGFloat(256), blue: CGFloat(256), alpha: 1)
+            centerView.layer.borderWidth = 7
+        }
+    }
+    
+    @IBOutlet weak var wheelControl: SwiftFortuneWheel! {
+        didSet {
+            wheelControl.configuration = .variousWheelSimpleConfiguration
+            wheelControl.slices = slices
+            wheelControl.pinImage = "whitePinArrow"
+        }
+    }
+    
+    var prizes = ["$30", "$10", "$250", "$20", "LOSE", "$5", "$500", "$80", "LOSE", "$200"]
+    
+    lazy var slices: [Slice] = {
+        let slices = prizes.map({ Slice.init(contents: [Slice.ContentType.text(text: $0, preferences: .variousWheelSimpleText)]) })
+        return slices
+    }()
+
+    var finishIndex: Int {
+        return Int.random(in: 0..<wheelControl.slices.count)
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+    
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        centerView.layer.cornerRadius = centerView.bounds.width / 2
+    }
+    
+    @IBAction func rotateTap(_ sender: Any) {
+        wheelControl.startAnimating(indefiniteRotationTimeInSeconds: 1, finishIndex: finishIndex) { (finished) in
+            print(finished)
+        }
+    }
+
+}